- изменены абстракции (удлаены, добавлены, изменены) - добавлены новые сервисы - добавлены новые репозитории

+++ src/main/java/application/services/FFmpeg/Resolution.java
@@ -7,7 +7,7 @@ public class Resolution {
 
     public Resolution(){}
 
-    public Resolution(int width, int height){
+    public Resolution(short width, short height){
         this.width = width;
         this.height = height;
     }
